% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubblechart.R
\name{bubblechart}
\alias{bubblechart}
\title{Bubble Chart}
\usage{
bubblechart(
  element_id,
  main_color,
  label_color,
  data = NULL,
  width = NULL,
  isInteractive = TRUE,
  height = "400px",
  activeColor = "transparent"
)
}
\description{
Nivo Bubble Chart
}
